doc_requirements.txt
sphinx
sphinx-rtd-theme
git+https://github.com/miyakogi/m2r.git
aafigure
docstr-coverage
sphinx-autoapi
sphinx-multiversion
sphinx-click
